+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: dahpark <dahpark@student.42seoul.k>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/10 12:02:29 by dahpark           #+#    #+#             */
-/*   Updated: 2020/10/12 18:58:09 by dahpark          ###   ########.fr       */
+/*   Updated: 2020/10/13 19:11:37 by dahpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
