+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: dahpark <dahpark@student.42seoul.k>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/09 15:01:49 by dahpark           #+#    #+#             */
-/*   Updated: 2020/10/12 18:58:52 by dahpark          ###   ########.fr       */
+/*   Updated: 2020/10/13 19:31:39 by dahpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
