+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: dahpark <dahpark@student.42seoul.k>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/07 15:39:12 by dahpark           #+#    #+#             */
-/*   Updated: 2020/10/12 19:01:32 by dahpark          ###   ########.fr       */
+/*   Updated: 2020/10/13 19:12:06 by dahpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' && i < n)
